Add no_namecoin_tls build tag
