+++ .travis/script
@@ -4,13 +4,21 @@
 #travis_retry go get github.com/tcnksm/ghr github.com/mitchellh/gox gopkg.in/alecthomas/gometalinter.v2 github.com/miekg/exdns/q github.com/namecoin/dns-prop279
 travis_retry go get github.com/tcnksm/ghr github.com/mitchellh/gox gopkg.in/alecthomas/gometalinter.v2
 
-go test -v ./...
+GOX_TAGS=""
+if [ "${NO_NAMECOIN_TLS}" = "1" ]; then
+  GOX_TAGS="no_namecoin_tls"
+fi
+
+go test -tags "${GOX_TAGS}" -v ./...
 RESULT=$?
 
 echo Done tests with exit code $RESULT
 if [ "$RESULT" != "0" ]; then
   return $RESULT
 fi
+
+# Static analysis for TLS mode will cover non-TLS mode as well.
+if [ "${NO_NAMECOIN_TLS}" != "1" ]; then
 
 # Static analysis
 
@@ -63,6 +71,8 @@ gometalinter.v2 --enable-all \
 ./...
 STATICRESULT2=$?
 
+fi
+
 # Test cross-compilation. The binaries produced are also used for release
 # upload in after_success if this is a release tag.
 
@@ -76,17 +86,19 @@ GOX_PARA=3
 REPOS="github.com/$TRAVIS_REPO_SLUG/..."
 
 # cgo crosscompile
-gox -parallel=$GOX_PARA -cgo -osarch 'linux/386 linux/amd64' -output "$GOPATH/releasing/idist/ncdns-$TRAVIS_TAG-{{.OS}}_{{.Arch}}/bin/{{.Dir}}" $REPOS
+gox -tags="$GOX_TAGS" -parallel=$GOX_PARA -cgo -osarch 'linux/386 linux/amd64' -output "$GOPATH/releasing/idist/ncdns-$TRAVIS_TAG-{{.OS}}_{{.Arch}}/bin/{{.Dir}}" $REPOS
 RESULT1=$?
 
 # non-cgo crosscompile
-gox -parallel=$GOX_PARA -osarch 'darwin/386 darwin/amd64 linux/arm linux/arm64 linux/ppc64 linux/ppc64le freebsd/386 freebsd/amd64 freebsd/arm openbsd/386 openbsd/amd64 netbsd/386 netbsd/amd64 netbsd/arm dragonfly/amd64 solaris/amd64 windows/386 windows/amd64' -output "$GOPATH/releasing/idist/ncdns-$TRAVIS_TAG-{{.OS}}_{{.Arch}}/bin/{{.Dir}}" $REPOS
+gox -tags="$GOX_TAGS" -parallel=$GOX_PARA -osarch 'darwin/386 darwin/amd64 linux/arm linux/arm64 linux/ppc64 linux/ppc64le freebsd/386 freebsd/amd64 freebsd/arm openbsd/386 openbsd/amd64 netbsd/386 netbsd/amd64 netbsd/arm dragonfly/amd64 solaris/amd64 windows/386 windows/amd64' -output "$GOPATH/releasing/idist/ncdns-$TRAVIS_TAG-{{.OS}}_{{.Arch}}/bin/{{.Dir}}" $REPOS
 RESULT2=$?
 
 echo cgo crosscompile exited with code $RESULT1
 echo non-cgo crosscompile exited with code $RESULT2
+if [ "${NO_NAMECOIN_TLS}" != "1" ]; then
 echo critical gometalinter exited with code $STATICRESULT1
 echo non-critical gometalinter exited with code $STATICRESULT2
+fi
 
 if [ "$RESULT1" != "0" ]; then
   return $RESULT1
@@ -94,6 +106,8 @@ fi
 if [ "$RESULT2" != "0" ]; then
   return $RESULT2
 fi
+if [ "${NO_NAMECOIN_TLS}" != "1" ]; then
 if [ "$STATICRESULT1" != "0" ]; then
   return $STATICRESULT1
+fi
 fi
