+++ backend/backend_notls.go
@@ -1,0 +1,7 @@
+// +build no_namecoin_tls
+
+package backend
+
+func lookupReadyError() error {
+	return nil
+}
