+++ ncdomain/convert_notls_test.go
@@ -1,0 +1,5 @@
+// +build no_namecoin_tls
+
+package ncdomain_test
+
+const tlsaDisabled = true
