+++ server/server.go
@@ -14,8 +14,6 @@ import (
 	"github.com/miekg/dns"
 	"github.com/namecoin/ncdns/backend"
 	"github.com/namecoin/ncdns/namecoin"
-	"github.com/namecoin/ncdns/tlsoverridefirefox/tlsoverridefirefoxsync"
-	"github.com/namecoin/tlsrestrictnss/tlsrestrictnsssync"
 	"gopkg.in/hlandau/madns.v1"
 )
 
@@ -218,17 +216,7 @@ func (s *Server) Start() error {
 	s.wgStart.Wait()
 	log.Info("Listeners started")
 
-	err := tlsoverridefirefoxsync.Start(s.namecoinConn, s.cfg.CanonicalSuffix)
-	if err != nil {
-		return fmt.Errorf("Couldn't start Firefox override sync: %s", err)
-	}
-
-	err = tlsrestrictnsssync.Start()
-	if err != nil {
-		return fmt.Errorf("Couldn't start tlsrestrictnss sync: %s", err)
-	}
-
-	return nil
+	return s.StartBackgroundTasks()
 }
 
 func (s *Server) doRunListener(ds *dns.Server) {
