+++ server/server_notls.go
@@ -1,0 +1,7 @@
+// +build no_namecoin_tls
+
+package server
+
+func (s *Server) StartBackgroundTasks() error {
+	return nil
+}
